Part one of ce 1
