+++ CodeExecutor/development.properties
@@ -1,0 +1,1 @@
+bootstrap.servers=localhost:9092
