+++ CodeExecutor/go.mod
@@ -1,5 +1,5 @@
-module awesomeProject
+module github.com/megamxl/escape-doom/CodeExecutor
 
-go 1.17
+go 1.23
 
 require github.com/confluentinc/confluent-kafka-go v1.9.2 // indirect
