+++ CodeExecutor/internal/constants/structs.go
@@ -1,0 +1,8 @@
+package constants
+
+type Request struct {
+	PlayerSessionId string `json:playerSessionId`
+	Language        string `json:language`
+	Code            string `json:code`
+	DateTime        string `json:dateTime`
+}
